clean code for readability (#11)
* separated into various files, added requirements

* separate social_capital and actions
